test-coffee: Fix the file modification test
The r variable was used instead of i to fill the buffer, resulting in
the end of the test loop after only a single iteration. The file was not
even closed at the end of each iteration although it is opened at the
beginning of each iteration, so the available file descriptors would
very quickly be exhausted.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau.dev@gmail.com>